use structs for producers and config
