+++ lib/beeline.ex
@@ -1,147 +1,6 @@
 defmodule Beeline do
-  @producer_schema [
-    adapter: [
-      doc: """
-      The adapter module to use for creating the producer. Use `:kelvin` for
-      EventStoreDB v3-5, `:volley` for EventStoreDB v20+, and `:dummy` for
-      test cases.
-      """,
-      type: {:in, [:kelvin, :volley, :dummy]},
-      required: true
-    ],
-    stream_name: [
-      doc: """
-      The name of the EventStoreDB stream to which this producer should
-      subscribe for events.
-      """,
-      type: :string,
-      required: true
-    ],
-    connection: [
-      doc: """
-      The module to use as a connection to form the subscription. When the
-      `:adapter` option is `:kelvin`, this should be an Extreme client module.
-      When the adapter is `:volley`, it should be a `Spear.Client` module.
-      """,
-      type: :atom,
-      required: true
-    ],
-    name: [
-      doc: """
-      The full GenServer name to use for this producer. When this option is
-      not provided, the name will be a formula of the name of the consumer
-      and the key in the keyword list of producers.
-      """,
-      type: :atom,
-      default: nil
-    ],
-    max_demand: [
-      doc: """
-      The maximum number of events the consumer is allowed to request from
-      this producer. This option can be configured to allow batch processing.
-      """,
-      type: :pos_integer,
-      default: 1
-    ],
-    min_demand: [
-      doc: """
-      The minimum number of events the consumer can request at a time from
-      this producer.
-      """,
-      type: :non_neg_integer,
-      default: 0
-    ]
-  ]
-
-  @schema [
-    name: [
-      doc: """
-      The GenServer name for the topology. The topology will build on this
-      name, using it as a prefix.
-      """,
-      type: :atom
-    ],
-    producers: [
-      doc: """
-      A list of producers to which the consumer should subscribe. See the
-      "producer options" section below for the schema.
-      """,
-      type: :keyword_list,
-      keys: [
-        *: [
-          type: :keyword_list,
-          keys:
-            Enum.map(@producer_schema, fn {k, v} ->
-              {k, put_in(v[:doc], false)}
-            end)
-        ]
-      ]
-    ],
-    get_stream_position: [
-      doc: """
-      A function to invoke in order to get the stream position for a producer.
-      This function should be a 1-arity function (anonymous or capture) where
-      the name of the producer is passed as the argument. This option may also
-      be passed as an MFA tuple where the producer name will be prepended to
-      the argument list. If this option is not provided, a default will be
-      fetched with `Application.fetch_env!(:beeline, :get_stream_position)`.
-      This configuration can be used to set a blanket function for all
-      beelines to use.
-      """,
-      type: {:or, [:mfa, {:fun, 1}]},
-      default: nil
-    ],
-    auto_subscribe?: [
-      doc: """
-      A function to invoke to determine whether each producer should
-      subscribe to events as it starts up. The argument passed is the
-      GenServer name of the producer. If this option is not provided,
-      a default will be fetched with
-      `Application.fetch_env!(:beeline, :auto_subscribe?)`.
-      """,
-      type: {:or, [:mfa, {:fun, 1}]},
-      default: nil
-    ],
-    subscribe_after: [
-      doc: """
-      A period in msec after initialization when each producer should
-      query the `:auto_subscribe?` function.
-      """,
-      type: {:or, [:mfa, :non_neg_integer]},
-      default: {Enum, :random, [3_000..5_000]}
-    ],
-    spawn_health_checkers?: [
-      doc: """
-      Controls whether the topology should spawn the HealthChecker children.
-      It can be useful to disable this in `Mix.env() in [:dev, :test]` as the
-      health checker provides little or no value in those environments and
-      can produce many log lines. If this option is left blank, it will be
-      gotten from the application environment defaulting to `true` with
-      `Application.get_env(:beeline, :spawn_health_checkers?, true)`.
-      """,
-      type: {:or, [:boolean, {:in, [nil]}]},
-      default: nil
-    ],
-    test_mode?: [
-      doc: """
-      Controls whether the topology should start up in test mode. In test
-      mode, any adapters set in producer specs are switched out with
-      the `:dummy` adapter. If this option is left blank, it will be
-      gotten from the application environment defaulting to `false` with
-      `Application.get_env(:beeline, :test_mode?, false)`.
-      """,
-      type: {:or, [:boolean, {:in, [nil]}]},
-      default: nil
-    ],
-    context: [
-      doc: """
-      A user-defined data structure which is used as the initial state of
-      the GenStage consumer process.
-      """,
-      type: :any,
-      default: nil
-    ]
-  ]
+  @schema Beeline.Config.schema()
+  @producer_schema Beeline.Producer.schema()
 
   @moduledoc """
   A tool for building in-order GenStage topologies for EventStoreDB
@@ -197,14 +56,14 @@ defmodule Beeline do
       use GenStage
 
       @impl GenStage
-      def init(opts) do
+      def init(config) do
         producers =
-          Enum.map(opts[:producers], fn {_key, producer} ->
-            {producer[:name],
-             Keyword.take(producer, [:max_demand, :min_demand])}
+          Enum.map(config.producers, fn {_key, producer} ->
+            {producer.name,
+             min_demand: producer.min_demand, max_demand: producer.max_demand}
           end)
 
-        {:consumer, opts[:context], subscribe_to: producers}
+        {:consumer, config.context, subscribe_to: producers}
       end
 
       defoverridable init: 1
@@ -262,76 +121,12 @@ defmodule Beeline do
                 reason.message
 
       {:ok, opts} ->
-        opts =
-          opts
-          |> Keyword.put(:module, module)
-          |> add_default_opts()
-
-        Beeline.Topology.start_link(opts)
+        opts
+        |> Keyword.put(:module, module)
+        |> Beeline.Config.source()
+        |> Beeline.Topology.start_link()
     end
   end
-
-  @doc false
-  def add_default_opts(opts) do
-    Enum.reduce(opts, [], &add_default_opt(&1, &2, opts))
-  end
-
-  @doc false
-  # coveralls-ignore-start
-  def add_default_opt({:get_stream_position, nil}, acc, _all_opts) do
-    get_stream_position = Application.fetch_env!(:beeline, :get_stream_position)
-
-    [{:get_stream_position, get_stream_position} | acc]
-  end
-
-  # coveralls-ignore-stop
-
-  def add_default_opt({:auto_subscribe?, nil}, acc, _all_opts) do
-    auto_subscribe? = Application.fetch_env!(:beeline, :auto_subscribe?)
-
-    [{:auto_subscribe?, auto_subscribe?} | acc]
-  end
-
-  def add_default_opt({:spawn_health_checkers?, nil}, acc, _all_opts) do
-    spawn_health_checkers? =
-      Application.get_env(:beeline, :spawn_health_checkers?, true)
-
-    [{:spawn_health_checkers?, spawn_health_checkers?} | acc]
-  end
-
-  def add_default_opt({:test_mode?, nil}, acc, _all_opts) do
-    test_mode? = Application.get_env(:beeline, :test_mode?, false)
-
-    [{:test_mode?, test_mode?} | acc]
-  end
-
-  def add_default_opt({:producers, producers}, acc, all_opts) do
-    producers =
-      producers
-      |> Enum.map(fn {key, producer} ->
-        producer =
-          Enum.reduce(
-            producer,
-            [],
-            &add_default_producer_opt(&1, &2, key, all_opts)
-          )
-
-        {key, producer}
-      end)
-
-    [{:producers, producers} | acc]
-  end
-
-  def add_default_opt({k, v}, acc, _all_opts), do: [{k, v} | acc]
-
-  @doc false
-  def add_default_producer_opt({:name, nil}, acc, key, all_opts) do
-    name = Module.concat(all_opts[:name], "Producer_#{key}")
-
-    [{:name, name} | acc]
-  end
-
-  def add_default_producer_opt({k, v}, acc, _key, _all_opts), do: [{k, v} | acc]
 
   @doc """
   Restarts the supervision tree of GenStages for the given Beeline topology
@@ -346,8 +141,10 @@ defmodule Beeline do
       :ok
   """
   @spec restart_stages(module()) :: :ok | {:error, term()}
-  def restart_stages(beeline) do
-    GenServer.call(beeline, :restart_stages)
+  def restart_stages(beeline) when is_atom(beeline) do
+    beeline
+    |> Module.concat(Topology)
+    |> GenServer.call(:restart_stages)
   end
 
   @doc """
@@ -418,21 +215,9 @@ defmodule Beeline do
   This function can be used to test running events through a topology.
   If there are multiple producers, one is picked at random.
   """
-  def test_events(events, beeline) do
-    GenServer.call(beeline, {:test_events, events})
+  def test_events(events, beeline) when is_atom(beeline) do
+    beeline
+    |> Module.concat(Topology)
+    |> GenServer.call({:test_events, events})
   end
-
-  @doc """
-  Returns the name of the consumer process given a beeline topology name
-
-  The consumer name is simply the concatenation of the topology's GenServer
-  name with "Consumer"
-
-  ## Examples
-
-      iex> Beeline.consumer(MyEventHandler)
-      MyEventHandler.Consumer
-  """
-  @spec consumer(module()) :: module()
-  def consumer(beeline), do: Module.concat(beeline, "Consumer")
 end
