+++ lib/beeline/topology.ex
@@ -7,21 +7,23 @@ defmodule Beeline.Topology do
 
   defstruct [:supervisor_pid, :config]
 
-  def start_link(opts) do
-    GenServer.start_link(__MODULE__, opts, Keyword.take(opts, [:name]))
+  def start_link(config) do
+    GenServer.start_link(__MODULE__, config,
+      name: Module.concat(config.name, "Topology")
+    )
   end
 
   @impl GenServer
-  def init(opts) do
-    {:ok, supervisor_pid} = spawn_supervisor(opts)
+  def init(config) do
+    {:ok, supervisor_pid} = spawn_supervisor(config)
 
-    {:ok, %__MODULE__{supervisor_pid: supervisor_pid, config: opts}}
+    {:ok, %__MODULE__{supervisor_pid: supervisor_pid, config: config}}
   end
 
   @impl GenServer
   def handle_call(:restart_stages, _from, state) do
     parent = state.supervisor_pid
-    target = Module.concat(state.config[:name], "StageSupervisor")
+    target = Module.concat(state.config.name, "StageSupervisor")
 
     spec =
       state.config
@@ -48,8 +50,8 @@ defmodule Beeline.Topology do
 
   def handle_call({:test_events, events}, _from, state) do
     producer =
-      state.config
-      |> get_in([:producers, Access.all(), Access.elem(1), :name])
+      state.config.producers
+      |> get_in([Access.all(), Access.elem(1), Access.key(:name)])
       |> Enum.random()
 
     events = Enum.map(events, &Beeline.as_subscription_event(&1, producer))
@@ -59,14 +61,14 @@ defmodule Beeline.Topology do
     {:reply, :ok, state}
   end
 
-  def spawn_supervisor(opts) do
+  def spawn_supervisor(config) do
     health_checkers =
-      if opts[:spawn_health_checkers?] do
-        opts[:producers]
+      if config.spawn_health_checkers? do
+        config.producers
         |> Enum.map(fn {_key, producer} ->
           {HealthChecker.StreamPosition,
-           event_listener: producer[:name],
-           get_current_stream_position: get_stream_position(opts, producer),
+           event_listener: producer.name,
+           get_current_stream_position: get_stream_position(config, producer),
            get_latest_stream_position: get_latest_stream_position(producer)}
         end)
       else
@@ -76,33 +78,33 @@ defmodule Beeline.Topology do
     children =
       health_checkers ++
         [
-          Supervisor.child_spec({StageSupervisor, opts},
-            id: StageSupervisor.name(opts)
+          Supervisor.child_spec({StageSupervisor, config},
+            id: StageSupervisor.name(config)
           )
         ]
 
     Supervisor.start_link(children,
       strategy: :one_for_one,
-      name: Module.concat(opts[:name], Supervisor)
+      name: Module.concat(config.name, Supervisor)
     )
   end
 
   @spec get_stream_position(Keyword.t(), Keyword.t()) ::
           (() -> non_neg_integer() | -1)
-  defp get_stream_position(opts, producer) do
-    case opts[:get_stream_position] do
+  defp get_stream_position(config, producer) do
+    case config.get_stream_position do
       {m, f, a} ->
-        fn -> apply(m, f, [producer[:name] | a]) end
+        fn -> apply(m, f, [producer.name | a]) end
 
       # coveralls-ignore-start
       function when is_function(function, 1) ->
-        fn -> function.(producer[:name]) end
+        fn -> function.(producer.name) end
 
       nil ->
         raise ArgumentError,
           message:
             "could not determine the " <>
-              "`:get_stream_position` function for Beeline #{inspect(opts[:name])}"
+              "`:get_stream_position` function for Beeline #{inspect(config.name)}"
 
         # coveralls-ignore-stop
     end
@@ -113,9 +115,9 @@ defmodule Beeline.Topology do
   defp get_latest_stream_position(producer) do
     fn ->
       Beeline.EventStoreDB.latest_event_number(
-        producer[:adapter],
-        producer[:connection],
-        producer[:stream_name]
+        producer.adapter,
+        producer.connection,
+        producer.stream_name
       )
     end
   end
